feat: add button for buy product
