+++ src/App.scss
@@ -96,3 +96,195 @@ img {
   border-radius: 8px;
   border: 0;
 }
+
+.modal-wrapper {
+  z-index: 99;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(13, 13, 13, 0.8);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: 1px solid blue;
+}
+
+.modal-content {
+  padding: 1%;
+  border: 1px solid blue;
+  border-radius: 0.5rem;
+  width: 40%;
+  background-color: #fff;
+  z-index: 99;
+  position: absolute;
+  background: rgb(0, 131, 201);
+}
+
+.modal-content__app {
+  width: 100%;
+  margin: auto;
+  display: block;
+  align-items: center;
+  justify-content: center;
+  height: fit-content;
+}
+
+.form-item {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.person-details__h2 {
+  text-align: center;
+  margin: 0 auto;
+  font-size: 90%;
+  color: rgb(0, 25, 38);
+}
+.form-item__input {
+  background: rgb(0, 73, 112);
+
+  width: 100%;
+  border: 1px solid rgb(0, 73, 112);
+  color: rgb(0, 18, 28);
+  border-radius: 2rem;
+  padding: 0.4rem;
+  outline: none;
+  font-size: 90%;
+  text-align: center;
+  margin: 2% auto;
+}
+
+.card-details {
+  margin-top: 1%;
+  width: 100%;
+}
+
+.card-details__h2 {
+  text-align: center;
+  margin: 0 auto;
+  font-size: 90%;
+  color: rgb(0, 25, 38);
+}
+.card-data {
+  border: 1px solid linear-gradient(90deg, rgb(0, 9, 36) 6%, rgb(9, 9, 121) 34%, rgb(0, 212, 255) 100%);
+  margin: 0 auto;
+  border-radius: 0.5rem;
+  background: rgb(0, 9, 36);
+  background: linear-gradient(90deg, rgb(0, 9, 36) 6%, rgb(9, 9, 121) 34%, rgb(0, 212, 255) 100%);
+  width: 55%;
+  height: 7rem;
+}
+
+.card-data__number {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  //margin: 7% auto 10%;
+  padding: 4% 0 0;
+}
+
+.card-data__img {
+  background-color: #fff;
+  width: 10%;
+  border-radius: 0.3rem;
+  padding: 0.3rem;
+}
+
+.card-data__input {
+  width: 80%;
+  background: rgba(0, 64, 255, 0.356);
+  border: 1px solid rgb(4, 52, 196);
+  border-radius: 0.3rem;
+  outline: none;
+  text-align: center;
+  margin-left: 2%;
+  font-size: 70%;
+  color: rgb(0, 18, 28);
+  font-weight: 600;
+  height: 1.3rem;
+}
+.valid-data__text {
+  font-size: 50%;
+  color: rgb(0, 205, 251);
+}
+
+.valid-data__input {
+  width: 60%;
+  height: 1rem;
+  background: rgb(0, 140, 255);
+  border: 1px solid blue;
+  border-radius: 0.3rem;
+  outline: none;
+  text-align: center;
+  margin-left: 2%;
+  font-size: 40%;
+  color: #000;
+}
+.valid-data {
+  width: 100%;
+  box-sizing: border-box;
+}
+.cvv-data {
+  width: 100%;
+  box-sizing: border-box;
+  padding-left: 5px;
+}
+
+.card-data__other-data {
+  margin-bottom: 5%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 90%;
+  margin: 2% auto 0;
+}
+
+.card-details_btn {
+  width: 40%;
+  font-size: 60%;
+  background-color: rgb(0, 47, 187);
+  border: 1px solid blue;
+  border-radius: 0.5rem;
+  padding: 1%;
+  cursor: pointer;
+  display: block;
+  margin: 5% auto;
+  color: black;
+}
+
+.ErrorInput {
+  color: red;
+  width: 100%;
+}
+.input-True {
+  background: green !important;
+}
+
+.ErrorInputCard {
+  color: red;
+  width: 100%;
+  height: 15px;
+  text-align: center;
+  font-size: 70%;
+}
+
+.intermediate {
+  height: 15px;
+  width: 100%;
+}
+
+.intermediate-2 {
+  margin-left: 17%;
+  height: 15px;
+  width: 65%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+.modal-wrapper__block {
+  display: none;
+}
