+++ src/components/ProductPage/ProductPage.scss
@@ -95,6 +95,20 @@
     margin-top: auto;
     display: flex;
     width: 100%;
-    justify-content: center;
+    justify-content: space-around;
+    button {
+      margin-top: auto;
+
+      a {
+        display: flex;
+        flex: 2;
+        align-items: center;
+        justify-content: center;
+        width: 100%;
+        height: 100%;
+        color: #fff;
+        text-decoration: none;
+      }
+    }
   }
 }
