+++ src/redux/slices/productsSlice.ts
@@ -4,10 +4,12 @@ import { ProductType } from "../../models/models";
 
 export interface ProductsState {
   products: ProductType[];
+  modal: boolean;
 }
 
 const initialState: ProductsState = {
   products: [],
+  modal: false,
 };
 
 export const productsSlice = createSlice({
@@ -17,9 +19,12 @@ export const productsSlice = createSlice({
     setProducts: (state, action: PayloadAction<ProductType[]>) => {
       state.products = action.payload;
     },
+    setModalState: (state, action: PayloadAction<boolean>) => {
+      state.modal = action.payload;
+    },
   },
 });
 
-export const { setProducts } = productsSlice.actions;
+export const { setProducts, setModalState } = productsSlice.actions;
 
 export default productsSlice.reducer;
